+++ Libplanet.Tests/Blocks/PreEvaluationBlockHeaderTest.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Immutable;
-using System.Linq;
 using System.Security.Cryptography;
 using Bencodex;
 using Libplanet.Blocks;
@@ -19,64 +18,12 @@ namespace Libplanet.Tests.Blocks
         protected readonly BlockContentFixture _contents;
         protected readonly Codec _codec;
         protected readonly HashAlgorithmType _hashAlgorithmType;
-        protected readonly (Nonce Nonce, ImmutableArray<byte> PreEvaluationHash) _validGenesisProof;
-        protected readonly (Nonce Nonce, ImmutableArray<byte> PreEvaluationHash) _validBlock1Proof;
-        protected readonly (Nonce Nonce, ImmutableArray<byte> PreEvaluationHash)
-            _invalidBlock1Proof;
 
         public PreEvaluationBlockHeaderTest()
         {
             _contents = new BlockContentFixture();
             _codec = new Codec();
             _hashAlgorithmType = BlockMetadata.HashAlgorithmType;
-
-            var validGenesisNonce = default(Nonce);
-            byte[] validGenesisBytes =
-                _codec.Encode(_contents.GenesisMetadata.MakeCandidateData(validGenesisNonce));
-            ImmutableArray<byte> validGenesisPreEvalHash =
-                _hashAlgorithmType.Digest(validGenesisBytes).ToImmutableArray();
-            _validGenesisProof = (validGenesisNonce, validGenesisPreEvalHash);
-
-            // Checks if the hard-coded proof of the fixture is up-to-date.  If it's outdated,
-            // throws an exception that prints a regenerated up-to-date one:
-            const int lastCheckedProtocolVersion = 4;
-            if (_contents.BlockMetadata1.ProtocolVersion > lastCheckedProtocolVersion)
-            {
-                (Nonce Nonce, ImmutableArray<byte> Digest) regen =
-                    Hashcash.Answer(
-                        n => new[] { _codec.Encode(_contents.BlockMetadata1.MakeCandidateData(n)) },
-                        _hashAlgorithmType,
-                        _contents.BlockMetadata1.Difficulty,
-                        0
-                    );
-                string nonceLit = string.Join(
-                    ", ",
-                    regen.Nonce.ByteArray.Select(b => b < 0x10 ? $"0x0{b:x}" : $"0x{b:x}")
-                );
-                throw new Exception(
-                    $"As the CurrentProtocolVersion was bumped from {lastCheckedProtocolVersion} " +
-                    $"to {BlockMetadata.CurrentProtocolVersion}, hard-coded nonce proofs in " +
-                    $"the fixture is now outdated.  Check {nameof(PreEvaluationBlockHeaderTest)} " +
-                    "constructor and update the hard-coded nonce to the following up-to-date one:" +
-                    $"\n\n    new {nameof(Nonce)}(new byte[] {{ {nonceLit} }})\n\n"
-                );
-            }
-
-            var validBlock1Nonce = new Nonce(
-                new byte[] { 0x1a, 0x0c, 0x46, 0x6f, 0x5d, 0x75, 0xe4, 0xd8, 0xad, 0x67 }
-            );
-            byte[] validBlock1Bytes =
-                _codec.Encode(_contents.BlockMetadata1.MakeCandidateData(validBlock1Nonce));
-            ImmutableArray<byte> validBlock1PreEvalHash =
-                _hashAlgorithmType.Digest(validBlock1Bytes).ToImmutableArray();
-            _validBlock1Proof = (validBlock1Nonce, validBlock1PreEvalHash);
-
-            var invalidBlock1Nonce = default(Nonce);
-            byte[] invalidBlock1Bytes =
-                _codec.Encode(_contents.BlockMetadata1.MakeCandidateData(invalidBlock1Nonce));
-            ImmutableArray<byte> invalidBlock1PreEvalHash =
-                _hashAlgorithmType.Digest(invalidBlock1Bytes).ToImmutableArray();
-            _invalidBlock1Proof = (invalidBlock1Nonce, invalidBlock1PreEvalHash);
         }
 
         [Fact]
@@ -133,9 +80,7 @@ namespace Libplanet.Tests.Blocks
                 LastCommit = invalidHeightLastCommit,
             };
             Assert.Throws<InvalidBlockLastCommitException>(
-                () => new PreEvaluationBlockHeader(
-                    metadata: invalidHeightMetadata,
-                    nonce: _validGenesisProof.Nonce));
+                () => new PreEvaluationBlockHeader(metadata: invalidHeightMetadata));
 
             // BlockHash of the last commit is invalid.
             var invalidBlockHashLastCommit = new BlockCommit(
@@ -157,9 +102,7 @@ namespace Libplanet.Tests.Blocks
                 LastCommit = invalidBlockHashLastCommit,
             };
             Assert.Throws<InvalidBlockLastCommitException>(
-                () => new PreEvaluationBlockHeader(
-                    metadata: invalidBlockHashMetadata,
-                    nonce: _validGenesisProof.Nonce));
+                () => new PreEvaluationBlockHeader(metadata: invalidBlockHashMetadata));
 
             // Some of the vote's signature are invalid.
             var invalidVoteSignatureLastCommit = new BlockCommit(
@@ -191,9 +134,7 @@ namespace Libplanet.Tests.Blocks
                 LastCommit = invalidVoteSignatureLastCommit,
             };
             Assert.Throws<InvalidBlockLastCommitException>(
-                () => new PreEvaluationBlockHeader(
-                    metadata: invalidVoteSignatureMetadata,
-                    nonce: _validGenesisProof.Nonce));
+                () => new PreEvaluationBlockHeader(metadata: invalidVoteSignatureMetadata));
 
             // Some of the vote's height are invalid.
             var invalidVoteHeightLastCommit = new BlockCommit(
@@ -225,9 +166,7 @@ namespace Libplanet.Tests.Blocks
                 LastCommit = invalidVoteHeightLastCommit,
             };
             Assert.Throws<InvalidBlockLastCommitException>(
-                () => new PreEvaluationBlockHeader(
-                    metadata: invalidVoteHeightMetadata,
-                    nonce: _validGenesisProof.Nonce));
+                () => new PreEvaluationBlockHeader(metadata: invalidVoteHeightMetadata));
 
             // Signature can be null for null or unknown votes.
             var validLastCommit = new BlockCommit(
@@ -265,19 +204,14 @@ namespace Libplanet.Tests.Blocks
                 ),
                 LastCommit = validLastCommit,
             };
-            _ = new PreEvaluationBlockHeader(
-                metadata: validMetadata,
-                nonce: _validGenesisProof.Nonce);
+            _ = new PreEvaluationBlockHeader(metadata: validMetadata);
         }
 
         [Fact]
         public void CopyConstructor()
         {
             BlockMetadata metadata = _contents.Genesis.Copy();
-            var preEvalBlock = new PreEvaluationBlockHeader(
-                metadata,
-                nonce: _validGenesisProof.Nonce
-            );
+            var preEvalBlock = new PreEvaluationBlockHeader(metadata);
             var copy = new PreEvaluationBlockHeader(preEvalBlock);
             AssertPreEvaluationBlockHeadersEqual(preEvalBlock, copy);
         }
@@ -290,18 +224,14 @@ namespace Libplanet.Tests.Blocks
             Bencodex.Types.Dictionary expectedGenesis = Bencodex.Types.Dictionary.Empty
                 .Add("index", 0L)
                 .Add("timestamp", "2021-09-06T04:46:39.123000Z")
-                .Add("nonce", _validGenesisProof.Nonce.ByteArray)
+                .Add("nonce", default(Nonce).ByteArray)
                 .Add(
                     "public_key",
                     ParseHex("0200e02709cc0c051dc105188c454a2e7ef7b36b85da34529d3abc1968167cf54f")
                 )
                 .Add("protocol_version", 4)
                 .Add("state_root_hash", default(HashDigest<SHA256>).ByteArray);
-            var genesis = new PreEvaluationBlockHeader(
-                _contents.GenesisMetadata,
-                nonce: _validGenesisProof.Nonce,
-                preEvaluationHash: _validGenesisProof.PreEvaluationHash
-            );
+            var genesis = new PreEvaluationBlockHeader(_contents.GenesisMetadata);
             AssertBencodexEqual(expectedGenesis, genesis.MakeCandidateData(default));
             HashDigest<SHA256> stateRootHash = random.NextHashDigest<SHA256>();
             AssertBencodexEqual(
@@ -312,7 +242,7 @@ namespace Libplanet.Tests.Blocks
             Bencodex.Types.Dictionary expectedBlock1 = Bencodex.Types.Dictionary.Empty
                 .Add("index", 1L)
                 .Add("timestamp", "2021-09-06T08:01:09.045000Z")
-                .Add("nonce", _validBlock1Proof.Nonce.ByteArray)
+                .Add("nonce", default(Nonce).ByteArray)
                 .Add(
                     "public_key",
                     ParseHex("0215ba27a461a986f4ce7bcda1fd73dc708da767d0405729edaacaad7b7ff60eed")
@@ -331,10 +261,7 @@ namespace Libplanet.Tests.Blocks
                 )
                 .Add("protocol_version", 4)
                 .Add("state_root_hash", default(HashDigest<SHA256>).ByteArray);
-            var block1 = new PreEvaluationBlockHeader(
-                _contents.BlockMetadata1,
-                nonce: _validBlock1Proof.Nonce
-            );
+            var block1 = new PreEvaluationBlockHeader(_contents.BlockMetadata1);
             AssertBencodexEqual(expectedBlock1, block1.MakeCandidateData(default));
             stateRootHash = random.NextHashDigest<SHA256>();
             AssertBencodexEqual(
@@ -346,7 +273,7 @@ namespace Libplanet.Tests.Blocks
             Bencodex.Types.Dictionary expectedBlockPv0 = Bencodex.Types.Dictionary.Empty
                 .Add("index", 0L)
                 .Add("timestamp", "2021-09-06T04:46:39.123000Z")
-                .Add("nonce", blockPv0.Nonce.ByteArray)
+                .Add("nonce", default(Nonce).ByteArray)
                 .Add("reward_beneficiary", ParseHex("268344BA46e6CA2A8a5096565548b9018bc687Ce"))
                 .Add("state_root_hash", default(HashDigest<SHA256>).ByteArray);
             AssertBencodexEqual(expectedBlockPv0, blockPv0.MakeCandidateData(default));
@@ -360,7 +287,7 @@ namespace Libplanet.Tests.Blocks
             Bencodex.Types.Dictionary expectedBlockPv1 = Bencodex.Types.Dictionary.Empty
                 .Add("index", 1L)
                 .Add("timestamp", "2021-09-06T08:01:09.045000Z")
-                .Add("nonce", blockPv1.Nonce.ByteArray)
+                .Add("nonce", default(Nonce).ByteArray)
                 .Add("reward_beneficiary", ParseHex("8a29de186B85560D708451101C4Bf02D63b25c50"))
                 .Add(
                     "previous_hash",
@@ -388,10 +315,7 @@ namespace Libplanet.Tests.Blocks
             );
 
             var key = _contents.Block1Key;
-            var block1 = new PreEvaluationBlockHeader(
-                _contents.BlockMetadata1,
-                nonce: _validBlock1Proof.Nonce
-            );
+            var block1 = new PreEvaluationBlockHeader(_contents.BlockMetadata1);
             ImmutableArray<byte> validSig = block1.MakeSignature(key, arbitraryHash);
             Assert.True(
                 key.PublicKey.Verify(
@@ -415,10 +339,7 @@ namespace Libplanet.Tests.Blocks
             Assert.Equal("privateKey", e.ParamName);
             Assert.Contains("does not match", e.Message);
 
-            var blockPv1 = new PreEvaluationBlockHeader(
-                _contents.BlockPv1,
-                _contents.BlockPv1.Mine().Nonce
-            );
+            var blockPv1 = new PreEvaluationBlockHeader(_contents.BlockPv1);
             InvalidOperationException e2 = Assert.Throws<InvalidOperationException>(
                 () => blockPv1.MakeSignature(key, arbitraryHash)
             );
@@ -433,17 +354,17 @@ namespace Libplanet.Tests.Blocks
                 "e6b3803208416556db8de50670aaf0b642e13c90afd77d24da8f642dc3e8f320"
             );
 
-            var block1 = new PreEvaluationBlockHeader(
-                _contents.BlockMetadata1,
-                nonce: _validBlock1Proof.Nonce
-            );
+            var block1 = new PreEvaluationBlockHeader(_contents.BlockMetadata1);
 
             // Same as block1.MakeSignature(_contents.Block1Key, arbitraryHash)
             ImmutableArray<byte> validSig = ByteUtil.ParseHex(
-                "304402205e1d796ee96a26b4479c3a8c4a9dc906dc8eb149dbe89c3c3e2fc7c85025022" +
-                "c022072da1febb0cc1f58c43fe7ce9f99460509f31e76f1875f0795d8610452e40c55"
+                "3044022100b285f684fe94524aa725c6b69cb858370f85af56420d275410e148b0ad18" +
+                "b3d9021f324c828b1dd949ebf73591bc0ac8858debae7c5aabc420fd4a1cec53d61e60"
             ).ToImmutableArray();
 
+            AssertBytesEqual(
+                block1.MakeSignature(_contents.Block1Key, arbitraryHash),
+                validSig);
             Assert.True(block1.VerifySignature(validSig, arbitraryHash));
             Assert.False(block1.VerifySignature(null, arbitraryHash));
             Assert.False(block1.VerifySignature(validSig, default));
@@ -454,10 +375,7 @@ namespace Libplanet.Tests.Blocks
                 )
             );
 
-            var blockPv1 = new PreEvaluationBlockHeader(
-                _contents.BlockPv1,
-                _contents.BlockPv1.Mine().Nonce
-            );
+            var blockPv1 = new PreEvaluationBlockHeader(_contents.BlockPv1);
             Assert.True(blockPv1.VerifySignature(null, arbitraryHash));
             Assert.False(blockPv1.VerifySignature(validSig, arbitraryHash));
         }
@@ -470,11 +388,7 @@ namespace Libplanet.Tests.Blocks
                 "e6b3803208416556db8de50670aaf0b642e13c90afd77d24da8f642dc3e8f320"
             );
 
-            var genesis = new PreEvaluationBlockHeader(
-                _contents.GenesisMetadata,
-                nonce: _validGenesisProof.Nonce,
-                preEvaluationHash: _validGenesisProof.PreEvaluationHash
-            );
+            var genesis = new PreEvaluationBlockHeader(_contents.GenesisMetadata);
             AssertBytesEqual(
                 fromHex("414d29402ca791f0540e34de7c27ed3728af440d262c3934b39fb3e9fd4644ee"),
                 genesis.DeriveBlockHash(default, null)
@@ -494,31 +408,26 @@ namespace Libplanet.Tests.Blocks
                 fromHex("3d358039543d29c872aa5b4e83dd497a2d91b430179c542107e063f8bdc3d55a"),
                 genesis.DeriveBlockHash(
                     arbitraryHash,
-                    genesis.MakeSignature(_contents.GenesisKey, arbitraryHash)
-                )
+                    genesis.MakeSignature(_contents.GenesisKey, arbitraryHash))
             );
 
-            var block1 = new PreEvaluationBlockHeader(
-                _contents.BlockMetadata1,
-                nonce: _validBlock1Proof.Nonce
-            );
+            var block1 = new PreEvaluationBlockHeader(_contents.BlockMetadata1);
             AssertBytesEqual(
-                fromHex("714cb54896cf865d4891d6687e8fa54631fb301c62a64b0f312217a029957235"),
+                fromHex("9c999d048603c32369bcd982ac2488f8d2782f339f0296ad537d4f039984dc47"),
                 block1.DeriveBlockHash(default, null)
             );
             AssertBytesEqual(
-                fromHex("e42aada1faf97d00042dfac3119913d724d757ab4f7bb510ce9c3a1109916677"),
+                fromHex("1ff0ca42037e91a07db6b42a2f0aadc7fc1249900033718d783334a90a333892"),
                 block1.DeriveBlockHash(default, block1.MakeSignature(_contents.Block1Key, default))
             );
             AssertBytesEqual(
-                fromHex("d8acdb1cd4087c9dafc6bb2e2c9d40fbd7b309f850176063dea394f84e9716d3"),
+                fromHex("5e63ed240742d7dac4b7f290f5a7afc51e2d410b13d2ad9690e56ac66486b23d"),
                 block1.DeriveBlockHash(arbitraryHash, null)
             );
             AssertBytesEqual(
-                fromHex("badf6772cac19b4b3c959acfa8799ee9a10ef6e23d27f0497dd294aa3f6be947"),
+                fromHex("54f7572223e4d642ca8772378528120392a9997655860a6ec108f3e4d8c1cc14"),
                 block1.DeriveBlockHash(
-                    arbitraryHash,
-                    block1.MakeSignature(_contents.Block1Key, arbitraryHash)
+                    arbitraryHash, block1.MakeSignature(_contents.Block1Key, arbitraryHash)
                 )
             );
         }