+++ Libplanet/Blocks/Block.cs
@@ -55,9 +55,7 @@ namespace Libplanet.Blocks
             : this(
                 new PreEvaluationBlock<T>(
                     new BlockContent<T>(header, transactions),
-                    header.Nonce,
-                    header.PreEvaluationHash
-                ),
+                    header.PreEvaluationHash),
                 header.StateRootHash,
                 header.Signature
             )
@@ -102,9 +100,6 @@ namespace Libplanet.Blocks
 
         /// <inheritdoc cref="IBlockMetadata.Index"/>
         public long Index => _preEvaluationBlock.Index;
-
-        /// <inheritdoc cref="IPreEvaluationBlockHeader.Nonce"/>
-        public Nonce Nonce => _preEvaluationBlock.Nonce;
 
         /// <inheritdoc cref="IBlockMetadata.Miner"/>
         public Address Miner => _preEvaluationBlock.Miner;