+++ CHANGES.md
@@ -16,6 +16,8 @@ To be released.
  -  Removed all total difficulty related parameters.  [[#PBFT]]
  -  Removed `IBlockPolicy.Difficulty` interface property and its
     implementations.  [[#PBFT]]
+ -  Removed `IPreEvaluationBlockHeader.Nonce` interface property and
+    its implementations.  [[#PBFT]]
  -  Removed `InvalidBlockTotalDifficultyException` class.  [[#PBFT]]
  -  Removed `InvalidBlockDifficultyException` class.  [[#PBFT]]